General overhaul and add vanillatweaks functionality
minecraft: Added vanillatweaks json files and mount the files as ro into the container. rcon password added so that the restarter can access rcon. Added create_console_in_pipe to fix issues with the restarter. Changed restart policy to unless-stopped.

mc-backup: Added initial delay to give time for rcon to become available.

restarter: Now uses rcon to properly initiate a graceful shutdown of the server. Added a 10 second warning to go along with the 5 minute warning. Restarter no longer restarts the container, rather the minecraft container was set to restart unless-stopped so that the server will come back up on it's own as soon as it's able to.